track network gas fees
